Added a new resource
Hey, I have added a reference link for Iframes in HTML under What is an iframe and how it works. I came upon this article while looking for resources to learn HTML. This citation, in my opinion, will enhance the content of this article. Hope that my contribution will benefit other learners.